+++ workspace/solutions/rust/src/s0/s0/s0/s1/mod.rs
@@ -1,0 +1,2 @@
+pub mod s_0001_00;
+pub mod s_0001_test;